Refactor: mypage와 관련된 DTO, VO 분리
