+++ ifive/src/main/java/kr/co/milionvolt/ifive/domain/userinfo/UserInfoReservationListVO.java
@@ -1,4 +1,4 @@
-package kr.co.milionvolt.ifive.domain.reservation;
+package kr.co.milionvolt.ifive.domain.userinfo;
 
 import lombok.*;
