+++ ifive/src/main/java/kr/co/milionvolt/ifive/service/user/UserServiceImpl.java
@@ -1,13 +1,13 @@
 package kr.co.milionvolt.ifive.service.user;
 
-import kr.co.milionvolt.ifive.domain.payment.UserInfoPaymentListVO;
-import kr.co.milionvolt.ifive.domain.reservation.UserInfoReservationListVO;
+import kr.co.milionvolt.ifive.domain.userinfo.UserInfoPaymentListVO;
+import kr.co.milionvolt.ifive.domain.userinfo.UserInfoReservationListVO;
 import kr.co.milionvolt.ifive.domain.user.PasswordDTO;
-import kr.co.milionvolt.ifive.domain.user.UserDashboradUserCarDTO;
-import kr.co.milionvolt.ifive.domain.user.UserInfoDTO;
-import kr.co.milionvolt.ifive.domain.usercar.CarBatteryAndChargerTypeUpdateDTO;
-import kr.co.milionvolt.ifive.domain.usercar.UserCarInfoDTO;
-import kr.co.milionvolt.ifive.domain.usercar.CarNumberAndModelUpdateDTO;
+import kr.co.milionvolt.ifive.domain.userinfo.UserDashboradUserCarDTO;
+import kr.co.milionvolt.ifive.domain.userinfo.UserInfoDTO;
+import kr.co.milionvolt.ifive.domain.userinfo.CarBatteryAndChargerTypeUpdateDTO;
+import kr.co.milionvolt.ifive.domain.userinfo.UserCarInfoDTO;
+import kr.co.milionvolt.ifive.domain.userinfo.CarNumberAndModelUpdateDTO;
 import kr.co.milionvolt.ifive.mapper.UserMapper;
 import kr.co.milionvolt.ifive.domain.user.UserVO;
 import lombok.RequiredArgsConstructor;
